# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/probes": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driving"
                ],
                "summary": "获取探头",
                "parameters": [
                    {
                        "type": "number",
                        "description": "当前位置经度",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "当前位置纬度",
                        "name": "lon",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "获取附近多少公里的探头，0为获取所有探头",
                        "name": "near",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiserver.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.ProbeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/route": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driving"
                ],
                "summary": "路线规划，获取需要避让的区域",
                "parameters": [
                    {
                        "type": "number",
                        "name": "from_lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "name": "from_lon",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "name": "to_lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "name": "to_lon",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回结果",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiserver.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/apis.RouteResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apis.DebugResp": {
            "type": "object",
            "properties": {
                "probe_count": {
                    "type": "integer"
                },
                "route_count": {
                    "type": "integer"
                },
                "routes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/apis.DebugRouteResp"
                    }
                }
            }
        },
        "apis.DebugRouteResp": {
            "type": "object",
            "properties": {
                "cur_to_next_to_probe": {
                    "description": "本次计算后需要避让的探头，格式为：cur;next;probe (A1;A2;探头)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "route_info": {
                    "description": "本次路线信息",
                    "type": "string"
                },
                "route_probe_info": {
                    "description": "本次路线信息传入的避让探头, 第一次传入的为空",
                    "type": "string"
                }
            }
        },
        "apis.ProbeResp": {
            "type": "object",
            "properties": {
                "probes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/drive.Coord"
                    }
                }
            }
        },
        "apis.RouteResp": {
            "type": "object",
            "properties": {
                "avoid_areas": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/drive.Coord"
                        }
                    }
                },
                "debug": {
                    "$ref": "#/definitions/apis.DebugResp"
                }
            }
        },
        "apiserver.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string"
                }
            }
        },
        "drive.Coord": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        }
    }
}